std types( uint8("def_types.h") -> uint8_t)

+++ config.h
@@ -9,9 +9,9 @@
 //__builtin_avr_delay_cycles
 //__delay_cycles
 
-#define wait_us( q)  __builtin_avr_delay_cycles( ( unsigned long)( ( q) * CLK_US + 0.50))
-#define wait_us2( q, diff)  __builtin_avr_delay_cycles( ( unsigned long)( ( q) * CLK_US + 0.50) - diff)
-#define wait_ms( q)  __builtin_avr_delay_cycles( ( unsigned long)( ( q) * 1000.0 * CLK_US))
+#define wait_us( q)  __builtin_avr_delay_cycles( ( uint32_t)( ( q) * CLK_US + 0.50))
+#define wait_us2( q, diff)  __builtin_avr_delay_cycles( ( uint32_t)( ( q) * CLK_US + 0.50) - diff)
+#define wait_ms( q)  __builtin_avr_delay_cycles( ( uint32_t)( ( q) * 1000.0 * CLK_US))
 
 ////////////////////////////////////////////////////////////////////////////////
 //#define LED_PIN  BIT0
